requirements.txt
Flask==2.2.3
Jinja2==3.1.2
Pillow==9.5.0
mysql-connector==2.2.9
mysqlclient==2.1.1
pytest==7.3.1
PyMySQL==1.0.3
cryptography==40.0.2
DateTime==5.1
